pyproject.toml
[tool.poetry]
name = "hw-mlops"
version = "0.1.0"
description = "MLOps homework"
authors = ["Nikolai Kochetkov <professor3222@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "hw_mlops"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
urllib3 = "1.26.15"
scikit-learn = "^1.3.2"
mlflow = "^2.9.2"
tensorboard = "^2.15.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
dvc = "^3.37.0"
dvc-gdrive = "^3.0.0"
ipykernel = "^6.28.0"
hydra-core = "^1.3.2"

[tool.black]
line-length = 90
target-version = ['py39']

[tool.flake8]
max-line-length = 90

[tool.isort]
line_length = 90

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
